Created Menu Scene and blank Game Scene
Menu scene now has three buttons, a "play" button that moves you to the next scene in the build index (which is the game scene currently), an "options" button which has a volume slider that controls the volume of the game music and a 'back' button to bring you back to the main menu, and a "quit" button which should quit the program eventually (right now it really just prints to console).

I also added 2 scripts, 'MainMenu' to deal with controlling the buttons and 'SoundManager' to give functionality to the volume slider.

There is also a 'Gold' color gradient in the assets folder for the main menu text and the sound file for the background music